src/request: split MPIR implementation from MPI functions
