+++ src/mpi/request/Makefile.mk
@@ -23,4 +23,5 @@ mpi_sources +=                           \
     src/mpi/request/waitsome.c
 
 mpi_core_sources += \
+    src/mpi/request/request_impl.c \
     src/mpi/request/mpir_request.c
