+++ src/mpi/request/cancel.c
@@ -22,98 +22,6 @@ int MPI_Cancel(MPI_Request * request) __attribute__ ((weak, alias("PMPI_Cancel")
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Cancel
 #define MPI_Cancel PMPI_Cancel
-
-
-int MPIR_Cancel(MPIR_Request * request_ptr)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    switch (request_ptr->kind) {
-        case MPIR_REQUEST_KIND__SEND:
-            {
-                mpi_errno = MPID_Cancel_send(request_ptr);
-                MPIR_ERR_CHECK(mpi_errno);
-                break;
-            }
-
-        case MPIR_REQUEST_KIND__RECV:
-            {
-                mpi_errno = MPID_Cancel_recv(request_ptr);
-                MPIR_ERR_CHECK(mpi_errno);
-                break;
-            }
-
-        case MPIR_REQUEST_KIND__PREQUEST_SEND:
-            {
-                if (request_ptr->u.persist.real_request != NULL) {
-                    if (request_ptr->u.persist.real_request->kind != MPIR_REQUEST_KIND__GREQUEST) {
-                        /* jratt@us.ibm.com: I don't know about the bsend
-                         * comment below, but the CC stuff on the next
-                         * line is *really* needed for persistent Bsend
-                         * request cancels.  The CC of the parent was
-                         * disconnected from the child to allow an
-                         * MPI_Wait in user-level to complete immediately
-                         * (mpid/dcmfd/src/persistent/mpid_startall.c).
-                         * However, if the user tries to cancel the parent
-                         * (and thereby cancels the child), we cannot just
-                         * say the request is done.  We need to re-link
-                         * the parent's cc_ptr to the child CC, thus
-                         * causing an MPI_Wait on the parent to block
-                         * until the child is canceled or completed.
-                         */
-                        request_ptr->cc_ptr = request_ptr->u.persist.real_request->cc_ptr;
-                        mpi_errno = MPID_Cancel_send(request_ptr->u.persist.real_request);
-                        MPIR_ERR_CHECK(mpi_errno);
-                    } else {
-                        /* This is needed for persistent Bsend requests */
-                        /* FIXME why do we directly access the partner request's
-                         * cc field?  shouldn't our cc_ptr be set to the address
-                         * of the partner req's cc field? */
-                        mpi_errno = MPIR_Grequest_cancel(request_ptr->u.persist.real_request,
-                                                         MPIR_cc_is_complete(&request_ptr->
-                                                                             u.persist.
-                                                                             real_request->cc));
-                        MPIR_ERR_CHECK(mpi_errno);
-                    }
-                } else {
-                    MPIR_ERR_SETANDJUMP(mpi_errno, MPI_ERR_REQUEST, "**requestpersistactive");
-                }
-                break;
-            }
-
-        case MPIR_REQUEST_KIND__PREQUEST_RECV:
-            {
-                if (request_ptr->u.persist.real_request != NULL) {
-                    mpi_errno = MPID_Cancel_recv(request_ptr->u.persist.real_request);
-                    MPIR_ERR_CHECK(mpi_errno);
-                } else {
-                    MPIR_ERR_SETANDJUMP(mpi_errno, MPI_ERR_REQUEST, "**requestpersistactive");
-                }
-                break;
-            }
-
-        case MPIR_REQUEST_KIND__GREQUEST:
-            {
-                mpi_errno =
-                    MPIR_Grequest_cancel(request_ptr, MPIR_cc_is_complete(&request_ptr->cc));
-                MPIR_ERR_CHECK(mpi_errno);
-                break;
-            }
-
-            /* --BEGIN ERROR HANDLING-- */
-        default:
-            {
-                MPIR_ERR_SETANDJUMP(mpi_errno, MPI_ERR_INTERN, "**cancelunknown");
-            }
-            /* --END ERROR HANDLING-- */
-    }
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-
-    goto fn_exit;
-}
 #endif
