+++ src/mpi/request/test.c
@@ -23,75 +23,6 @@ int MPI_Test(MPI_Request * request, int *flag, MPI_Status * status)
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Test
 #define MPI_Test PMPI_Test
-
-int MPIR_Test_state(MPIR_Request * request_ptr, int *flag, MPI_Status * status,
-                    MPID_Progress_state * state)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    mpi_errno = MPID_Progress_test(state);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (MPIR_Request_has_poll_fn(request_ptr)) {
-        mpi_errno = MPIR_Grequest_poll(request_ptr, status);
-        MPIR_ERR_CHECK(mpi_errno);
-    }
-
-    if (MPIR_Request_is_complete(request_ptr))
-        *flag = TRUE;
-    else
-        *flag = FALSE;
-
-  fn_exit:
-    return mpi_errno;
-
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Test_impl(MPIR_Request * request_ptr, int *flag, MPI_Status * status)
-{
-    return MPIR_Test_state(request_ptr, flag, status, NULL);
-}
-
-int MPIR_Test(MPI_Request * request, int *flag, MPI_Status * status)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Request *request_ptr = NULL;
-
-    /* If this is a null request handle, then return an empty status */
-    if (*request == MPI_REQUEST_NULL) {
-        MPIR_Status_set_empty(status);
-        *flag = TRUE;
-        goto fn_exit;
-    }
-
-    MPIR_Request_get_ptr(*request, request_ptr);
-    MPIR_Assert(request_ptr != NULL);
-
-    mpi_errno = MPID_Test(request_ptr, flag, status);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (*flag) {
-        mpi_errno = MPIR_Request_completion_processing(request_ptr, status);
-        if (!MPIR_Request_is_persistent(request_ptr)) {
-            MPIR_Request_free(request_ptr);
-            *request = MPI_REQUEST_NULL;
-        }
-        MPIR_ERR_CHECK(mpi_errno);
-        /* Fall through to the exit */
-    } else if (unlikely(MPIR_Request_is_anysrc_mismatched(request_ptr))) {
-        MPIR_ERR_SET(mpi_errno, MPIX_ERR_PROC_FAILED_PENDING, "**failure_pending");
-        if (status != MPI_STATUS_IGNORE)
-            status->MPI_ERROR = mpi_errno;
-        goto fn_fail;
-    }
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif
 
 /*@
