+++ src/mpi/request/testall.c
@@ -5,10 +5,6 @@
 
 #include "mpiimpl.h"
 
-#if !defined(MPIR_REQUEST_PTR_ARRAY_SIZE)
-#define MPIR_REQUEST_PTR_ARRAY_SIZE 16
-#endif
-
 /* -- Begin Profiling Symbol Block for routine MPI_Testall */
 #if defined(HAVE_PRAGMA_WEAK)
 #pragma weak MPI_Testall = PMPI_Testall
@@ -27,235 +23,6 @@ int MPI_Testall(int count, MPI_Request array_of_requests[], int *flag,
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Testall
 #define MPI_Testall PMPI_Testall
-
-
-int MPIR_Testall_state(int count, MPIR_Request * request_ptrs[], int *flag,
-                       MPI_Status array_of_statuses[], int requests_property,
-                       MPID_Progress_state * state)
-{
-    int i;
-    int mpi_errno = MPI_SUCCESS;
-    int n_completed = 0;
-
-    mpi_errno = MPID_Progress_test(state);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    if (requests_property & MPIR_REQUESTS_PROPERTY__NO_GREQUESTS) {
-        for (i = 0; i < count; i++) {
-            if ((i + 1) % MPIR_CVAR_REQUEST_POLL_FREQ == 0) {
-                mpi_errno = MPID_Progress_test(state);
-                MPIR_ERR_CHECK(mpi_errno);
-            }
-
-            if (request_ptrs[i] == NULL || MPIR_Request_is_complete(request_ptrs[i])) {
-                n_completed++;
-            } else {
-                break;
-            }
-        }
-    } else {
-        for (i = 0; i < count; i++) {
-            if ((i + 1) % MPIR_CVAR_REQUEST_POLL_FREQ == 0) {
-                mpi_errno = MPID_Progress_test(state);
-                MPIR_ERR_CHECK(mpi_errno);
-            }
-
-            if (request_ptrs[i] != NULL) {
-                if (MPIR_Request_has_poll_fn(request_ptrs[i])) {
-                    mpi_errno = MPIR_Grequest_poll(request_ptrs[i], &array_of_statuses[i]);
-                    MPIR_ERR_CHECK(mpi_errno);
-                }
-                if (MPIR_Request_is_complete(request_ptrs[i])) {
-                    n_completed++;
-                }
-            } else {
-                n_completed++;
-            }
-        }
-    }
-    *flag = (n_completed == count) ? TRUE : FALSE;
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Testall_impl(int count, MPIR_Request * request_ptrs[], int *flag,
-                      MPI_Status array_of_statuses[], int requests_property)
-{
-    return MPIR_Testall_state(count, request_ptrs, flag, array_of_statuses,
-                              requests_property, NULL);
-}
-
-int MPIR_Testall(int count, MPI_Request array_of_requests[], int *flag,
-                 MPI_Status array_of_statuses[])
-{
-    MPIR_Request *request_ptr_array[MPIR_REQUEST_PTR_ARRAY_SIZE];
-    MPIR_Request **request_ptrs = request_ptr_array;
-    int i;
-    int n_completed;
-    int active_flag;
-    int rc = MPI_SUCCESS;
-    int proc_failure = FALSE;
-    int mpi_errno = MPI_SUCCESS;
-    int requests_property = MPIR_REQUESTS_PROPERTY__OPT_ALL;
-    int ignoring_status = (array_of_statuses == MPI_STATUSES_IGNORE);
-    MPIR_CHKLMEM_DECL(1);
-
-    int ii, icount, impi_errno;
-    n_completed = 0;
-
-    /* Convert MPI request handles to a request object pointers */
-    if (count > MPIR_REQUEST_PTR_ARRAY_SIZE) {
-        MPIR_CHKLMEM_MALLOC_ORJUMP(request_ptrs, MPIR_Request **,
-                                   count * sizeof(MPIR_Request *), mpi_errno, "request pointers",
-                                   MPL_MEM_OBJECT);
-    }
-
-    for (ii = 0; ii < count; ii += MPIR_CVAR_REQUEST_BATCH_SIZE) {
-        icount = count - ii > MPIR_CVAR_REQUEST_BATCH_SIZE ?
-            MPIR_CVAR_REQUEST_BATCH_SIZE : count - ii;
-
-        requests_property = MPIR_REQUESTS_PROPERTY__OPT_ALL;
-
-        for (i = ii; i < ii + icount; i++) {
-            if (array_of_requests[i] != MPI_REQUEST_NULL) {
-                MPIR_Request_get_ptr(array_of_requests[i], request_ptrs[i]);
-                /* Validate object pointers if error checking is enabled */
-#ifdef HAVE_ERROR_CHECKING
-                {
-                    MPID_BEGIN_ERROR_CHECKS;
-                    {
-                        MPIR_Request_valid_ptr(request_ptrs[i], mpi_errno);
-                        if (mpi_errno)
-                            goto fn_fail;
-                    }
-                    MPID_END_ERROR_CHECKS;
-                }
-#endif
-                if (request_ptrs[i]->kind != MPIR_REQUEST_KIND__RECV &&
-                    request_ptrs[i]->kind != MPIR_REQUEST_KIND__SEND) {
-                    requests_property &= ~MPIR_REQUESTS_PROPERTY__SEND_RECV_ONLY;
-                    if (request_ptrs[i]->kind == MPIR_REQUEST_KIND__GREQUEST) {
-                        requests_property &= ~MPIR_REQUESTS_PROPERTY__NO_GREQUESTS;
-                    }
-                }
-            } else {
-                request_ptrs[i] = NULL;
-                requests_property &= ~MPIR_REQUESTS_PROPERTY__NO_NULL;
-            }
-        }
-
-        impi_errno = MPID_Testall(icount, &request_ptrs[ii], flag,
-                                  ignoring_status ? MPI_STATUSES_IGNORE : &array_of_statuses[ii],
-                                  requests_property);
-        if (impi_errno != MPI_SUCCESS) {
-            mpi_errno = impi_errno;
-            goto fn_fail;
-        }
-
-        for (i = ii; i < ii + icount; i++) {
-            if (request_ptrs[i] == NULL || MPIR_Request_is_complete(request_ptrs[i]))
-                n_completed += 1;
-#ifdef HAVE_ERROR_CHECKING
-            {
-                MPID_BEGIN_ERROR_CHECKS;
-                {
-                    if (request_ptrs[i] == NULL)
-                        continue;
-                    if (MPIR_Request_is_complete(request_ptrs[i])) {
-                        rc = MPIR_Request_get_error(request_ptrs[i]);
-                        if (rc != MPI_SUCCESS) {
-                            if (MPIX_ERR_PROC_FAILED == MPIR_ERR_GET_CLASS(rc) ||
-                                MPIX_ERR_PROC_FAILED_PENDING == MPIR_ERR_GET_CLASS(rc))
-                                proc_failure = TRUE;
-                            mpi_errno = MPI_ERR_IN_STATUS;
-                        }
-                    } else if (unlikely(MPIR_Request_is_anysrc_mismatched(request_ptrs[i]))) {
-                        mpi_errno = MPI_ERR_IN_STATUS;
-                        MPIR_ERR_SET(rc, MPIX_ERR_PROC_FAILED_PENDING, "**failure_pending");
-                        if (!ignoring_status)
-                            array_of_statuses[i].MPI_ERROR = rc;
-                        proc_failure = TRUE;
-                    }
-                }
-                MPID_END_ERROR_CHECKS;
-            }
-#endif
-        }
-    }
-
-    *flag = (n_completed == count) ? TRUE : FALSE;
-
-    /* We only process completion of requests if all are finished, or
-     * there is an error. */
-    if (!(*flag || mpi_errno == MPI_ERR_IN_STATUS))
-        goto fn_exit;
-
-    if (ignoring_status && (requests_property & MPIR_REQUESTS_PROPERTY__SEND_RECV_ONLY)) {
-        for (i = 0; i < count; i++) {
-            if (request_ptrs[i] != NULL && MPIR_Request_is_complete(request_ptrs[i]))
-                MPIR_Request_completion_processing_fastpath(&array_of_requests[i], request_ptrs[i]);
-        }
-        goto fn_exit;
-    }
-
-    if (ignoring_status) {
-        for (i = 0; i < count; i++) {
-            if (request_ptrs[i] != NULL && MPIR_Request_is_complete(request_ptrs[i])) {
-                MPIR_Request_completion_processing(request_ptrs[i], MPI_STATUS_IGNORE);
-                if (!MPIR_Request_is_persistent(request_ptrs[i])) {
-                    MPIR_Request_free(request_ptrs[i]);
-                    array_of_requests[i] = MPI_REQUEST_NULL;
-                }
-            }
-        }
-        goto fn_exit;
-    }
-
-    for (i = 0; i < count; i++) {
-        if (request_ptrs[i] != NULL) {
-            if (MPIR_Request_is_complete(request_ptrs[i])) {
-                active_flag = MPIR_Request_is_active(request_ptrs[i]);
-                rc = MPIR_Request_completion_processing(request_ptrs[i], &array_of_statuses[i]);
-                if (!MPIR_Request_is_persistent(request_ptrs[i])) {
-                    MPIR_Request_free(request_ptrs[i]);
-                    array_of_requests[i] = MPI_REQUEST_NULL;
-                }
-                if (mpi_errno == MPI_ERR_IN_STATUS) {
-                    if (active_flag) {
-                        array_of_statuses[i].MPI_ERROR = rc;
-                    } else {
-                        array_of_statuses[i].MPI_ERROR = MPI_SUCCESS;
-                    }
-                }
-            } else {
-                if (mpi_errno == MPI_ERR_IN_STATUS) {
-                    if (!proc_failure)
-                        array_of_statuses[i].MPI_ERROR = MPI_ERR_PENDING;
-                    else
-                        array_of_statuses[i].MPI_ERROR = MPIX_ERR_PROC_FAILED_PENDING;
-                }
-            }
-        } else {
-            MPIR_Status_set_empty(&array_of_statuses[i]);
-            if (mpi_errno == MPI_ERR_IN_STATUS) {
-                array_of_statuses[i].MPI_ERROR = MPI_SUCCESS;
-            }
-        }
-    }
-
-  fn_exit:
-    if (count > MPIR_REQUEST_PTR_ARRAY_SIZE) {
-        MPIR_CHKLMEM_FREEALL();
-    }
-
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif
 
 /*@
