+++ src/mpi/request/wait.c
@@ -22,93 +22,6 @@ int MPI_Wait(MPI_Request * request, MPI_Status * status) __attribute__ ((weak, a
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Wait
 #define MPI_Wait PMPI_Wait
-
-/* MPID_Wait call MPIR_Wait_state with initialized progress state */
-int MPIR_Wait_state(MPIR_Request * request_ptr, MPI_Status * status, MPID_Progress_state * state)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    while (!MPIR_Request_is_complete(request_ptr)) {
-        mpi_errno = MPID_Progress_wait(state);
-        MPIR_ERR_CHECK(mpi_errno);
-
-        if (unlikely(MPIR_Request_is_anysrc_mismatched(request_ptr))) {
-            mpi_errno = MPIR_Request_handle_proc_failed(request_ptr);
-            goto fn_fail;
-        }
-    }
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-/* legacy interface (for ch3) */
-int MPIR_Wait_impl(MPIR_Request * request_ptr, MPI_Status * status)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPID_Progress_state progress_state;
-
-    MPIR_Assert(request_ptr != NULL);
-    MPID_Progress_start(&progress_state);
-    mpi_errno = MPIR_Wait_state(request_ptr, status, &progress_state);
-    MPID_Progress_end(&progress_state);
-
-    return mpi_errno;
-}
-
-int MPIR_Wait(MPI_Request * request, MPI_Status * status)
-{
-    int mpi_errno = MPI_SUCCESS;
-    int active_flag;
-    MPIR_Request *request_ptr = NULL;
-
-    /* If this is a null request handle, then return an empty status */
-    if (*request == MPI_REQUEST_NULL) {
-        MPIR_Status_set_empty(status);
-        goto fn_exit;
-    }
-
-    MPIR_Request_get_ptr(*request, request_ptr);
-    MPIR_Assert(request_ptr != NULL);
-
-    if (!MPIR_Request_is_complete(request_ptr)) {
-        /* If this is an anysource request including a communicator with
-         * anysource disabled, convert the call to an MPI_Test instead so we
-         * don't get stuck in the progress engine. */
-        if (unlikely(MPIR_Request_is_anysrc_mismatched(request_ptr))) {
-            mpi_errno = MPIR_Test(request, &active_flag, status);
-            goto fn_exit;
-        }
-
-        if (MPIR_Request_has_poll_fn(request_ptr)) {
-            while (!MPIR_Request_is_complete(request_ptr)) {
-                mpi_errno = MPIR_Grequest_poll(request_ptr, status);
-                MPIR_ERR_CHECK(mpi_errno);
-
-                /* Avoid blocking other threads since I am inside an infinite loop */
-                MPID_THREAD_CS_YIELD(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-            }
-        } else {
-            mpi_errno = MPID_Wait(request_ptr, status);
-            MPIR_ERR_CHECK(mpi_errno);
-        }
-    }
-
-    mpi_errno = MPIR_Request_completion_processing(request_ptr, status);
-    if (!MPIR_Request_is_persistent(request_ptr)) {
-        MPIR_Request_free(request_ptr);
-        *request = MPI_REQUEST_NULL;
-    }
-    MPIR_ERR_CHECK(mpi_errno);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif
 
 /*@
