+++ src/mpi/request/waitall.c
@@ -5,11 +5,6 @@
 
 #include "mpiimpl.h"
 
-#if !defined(MPIR_REQUEST_PTR_ARRAY_SIZE)
-/* use a larger default size of 64 in order to enhance SQMR performance */
-#define MPIR_REQUEST_PTR_ARRAY_SIZE 64
-#endif
-
 /* -- Begin Profiling Symbol Block for routine MPI_Waitall */
 #if defined(HAVE_PRAGMA_WEAK)
 #pragma weak MPI_Waitall = PMPI_Waitall
@@ -28,223 +23,6 @@ int MPI_Waitall(int count, MPI_Request array_of_requests[], MPI_Status array_of_
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Waitall
 #define MPI_Waitall PMPI_Waitall
-
-/* MPID_Waitall call MPIR_Waitall_state with initialized progress state */
-int MPIR_Waitall_state(int count, MPIR_Request * request_ptrs[], MPI_Status array_of_statuses[],
-                       int requests_property, MPID_Progress_state * state)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    if (requests_property & MPIR_REQUESTS_PROPERTY__NO_NULL) {
-        for (int i = 0; i < count; ++i) {
-            while (!MPIR_Request_is_complete(request_ptrs[i])) {
-                mpi_errno = MPID_Progress_wait(state);
-                MPIR_ERR_CHECK(mpi_errno);
-            }
-        }
-    } else {
-        for (int i = 0; i < count; i++) {
-            if (request_ptrs[i] == NULL) {
-                continue;
-            }
-            /* wait for ith request to complete */
-            while (!MPIR_Request_is_complete(request_ptrs[i])) {
-                /* generalized requests should already be finished */
-                MPIR_Assert(request_ptrs[i]->kind != MPIR_REQUEST_KIND__GREQUEST);
-
-                mpi_errno = MPID_Progress_wait(state);
-                MPIR_ERR_CHECK(mpi_errno);
-            }
-        }
-    }
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-/* legacy interface (for ch3) */
-int MPIR_Waitall_impl(int count, MPIR_Request * request_ptrs[], MPI_Status array_of_statuses[],
-                      int requests_property)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPID_Progress_state progress_state;
-
-    MPID_Progress_start(&progress_state);
-    mpi_errno = MPIR_Waitall_state(count, request_ptrs, array_of_statuses, requests_property,
-                                   &progress_state);
-    MPID_Progress_end(&progress_state);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-int MPIR_Waitall(int count, MPI_Request array_of_requests[], MPI_Status array_of_statuses[])
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPIR_Request *request_ptr_array[MPIR_REQUEST_PTR_ARRAY_SIZE];
-    MPIR_Request **request_ptrs = request_ptr_array;
-    int i, j, ii, icount;
-    int n_completed;
-    int rc = MPI_SUCCESS;
-    int disabled_anysource = FALSE;
-    const int ignoring_statuses = (array_of_statuses == MPI_STATUSES_IGNORE);
-    int requests_property = MPIR_REQUESTS_PROPERTY__OPT_ALL;
-    MPIR_CHKLMEM_DECL(1);
-
-    /* Convert MPI request handles to a request object pointers */
-    if (count > MPIR_REQUEST_PTR_ARRAY_SIZE) {
-        MPIR_CHKLMEM_MALLOC(request_ptrs, MPIR_Request **, count * sizeof(MPIR_Request *),
-                            mpi_errno, "request pointers", MPL_MEM_OBJECT);
-    }
-
-    for (ii = 0; ii < count; ii += MPIR_CVAR_REQUEST_BATCH_SIZE) {
-        icount = count - ii > MPIR_CVAR_REQUEST_BATCH_SIZE ?
-            MPIR_CVAR_REQUEST_BATCH_SIZE : count - ii;
-
-        n_completed = 0;
-        requests_property = MPIR_REQUESTS_PROPERTY__OPT_ALL;
-
-        for (i = ii; i < ii + icount; i++) {
-            if (array_of_requests[i] != MPI_REQUEST_NULL) {
-                MPIR_Request_get_ptr(array_of_requests[i], request_ptrs[i]);
-                /* Validate object pointers if error checking is enabled */
-#ifdef HAVE_ERROR_CHECKING
-                {
-                    MPID_BEGIN_ERROR_CHECKS;
-                    {
-                        MPIR_Request_valid_ptr(request_ptrs[i], mpi_errno);
-                        MPIR_ERR_CHECK(mpi_errno);
-                        MPIR_ERR_CHKANDJUMP1((request_ptrs[i]->kind == MPIR_REQUEST_KIND__MPROBE),
-                                             mpi_errno, MPI_ERR_ARG, "**msgnotreq",
-                                             "**msgnotreq %d", i);
-                    }
-                    MPID_END_ERROR_CHECKS;
-                }
-#endif
-                /* If one of the requests is an anysource on a communicator that's
-                 * disabled such communication, convert this operation to a testall
-                 * instead to prevent getting stuck in the progress engine. */
-                if (unlikely(MPIR_Request_is_anysrc_mismatched(request_ptrs[i]))) {
-                    disabled_anysource = TRUE;
-                }
-
-                if (request_ptrs[i]->kind != MPIR_REQUEST_KIND__RECV &&
-                    request_ptrs[i]->kind != MPIR_REQUEST_KIND__SEND) {
-                    requests_property &= ~MPIR_REQUESTS_PROPERTY__SEND_RECV_ONLY;
-
-                    /* If this is extended generalized request, we can complete it here. */
-                    if (MPIR_Request_has_wait_fn(request_ptrs[i])) {
-                        while (!MPIR_Request_is_complete(request_ptrs[i]))
-                            MPIR_Grequest_wait(request_ptrs[i], &array_of_statuses[i]);
-                    }
-                }
-            } else {
-                if (!ignoring_statuses)
-                    MPIR_Status_set_empty(&array_of_statuses[i]);
-                request_ptrs[i] = NULL;
-                n_completed += 1;
-                requests_property &= ~MPIR_REQUESTS_PROPERTY__NO_NULL;
-            }
-        }
-
-        if (n_completed == icount) {
-            continue;
-        }
-
-        if (unlikely(disabled_anysource)) {
-            mpi_errno =
-                MPIR_Testall(count, array_of_requests, &disabled_anysource, array_of_statuses);
-            goto fn_exit;
-        }
-
-        mpi_errno = MPID_Waitall(icount, &request_ptrs[ii], array_of_statuses, requests_property);
-        MPIR_ERR_CHECK(mpi_errno);
-
-        if (requests_property == MPIR_REQUESTS_PROPERTY__OPT_ALL && ignoring_statuses) {
-            /* NOTE-O1: high-message-rate optimization.  For simple send and recv
-             * operations and MPI_STATUSES_IGNORE we use a fastpath approach that strips
-             * out as many unnecessary jumps and error handling as possible.
-             *
-             * Possible variation: permit request_ptrs[i]==NULL at the cost of an
-             * additional branch inside the for-loop below. */
-            for (i = ii; i < ii + icount; ++i) {
-                rc = MPIR_Request_completion_processing_fastpath(&array_of_requests[i],
-                                                                 request_ptrs[i]);
-                if (rc != MPI_SUCCESS) {
-                    MPIR_ERR_SET(mpi_errno, MPI_ERR_IN_STATUS, "**instatus");
-                    goto fn_exit;
-                }
-            }
-            continue;
-        }
-
-        if (ignoring_statuses) {
-            for (i = ii; i < ii + icount; i++) {
-                if (request_ptrs[i] == NULL)
-                    continue;
-                rc = MPIR_Request_completion_processing(request_ptrs[i], MPI_STATUS_IGNORE);
-                if (!MPIR_Request_is_persistent(request_ptrs[i])) {
-                    MPIR_Request_free(request_ptrs[i]);
-                    array_of_requests[i] = MPI_REQUEST_NULL;
-                }
-                if (rc != MPI_SUCCESS) {
-                    MPIR_ERR_SET(mpi_errno, MPI_ERR_IN_STATUS, "**instatus");
-                    goto fn_exit;
-                }
-            }
-            continue;
-        }
-
-        for (i = ii; i < ii + icount; i++) {
-            if (request_ptrs[i] == NULL)
-                continue;
-            rc = MPIR_Request_completion_processing(request_ptrs[i], &array_of_statuses[i]);
-            if (!MPIR_Request_is_persistent(request_ptrs[i])) {
-                MPIR_Request_free(request_ptrs[i]);
-                array_of_requests[i] = MPI_REQUEST_NULL;
-            }
-
-            if (rc == MPI_SUCCESS) {
-                array_of_statuses[i].MPI_ERROR = MPI_SUCCESS;
-            } else {
-                /* req completed with an error */
-                MPIR_ERR_SET(mpi_errno, MPI_ERR_IN_STATUS, "**instatus");
-
-                /* set the error code for this request */
-                array_of_statuses[i].MPI_ERROR = rc;
-
-                if (unlikely(MPIX_ERR_PROC_FAILED == MPIR_ERR_GET_CLASS(rc)))
-                    rc = MPIX_ERR_PROC_FAILED_PENDING;
-                else
-                    rc = MPI_ERR_PENDING;
-
-                /* set the error codes for the rest of the uncompleted requests to PENDING */
-                for (j = i + 1; j < count; ++j) {
-                    if (request_ptrs[j] == NULL) {
-                        /* either the user specified MPI_REQUEST_NULL, or this is a completed greq */
-                        array_of_statuses[j].MPI_ERROR = MPI_SUCCESS;
-                    } else {
-                        array_of_statuses[j].MPI_ERROR = rc;
-                    }
-                }
-                goto fn_exit;
-            }
-        }
-    }
-
-  fn_exit:
-    if (count > MPIR_REQUEST_PTR_ARRAY_SIZE) {
-        MPIR_CHKLMEM_FREEALL();
-    }
-
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
 #endif
 
 /*@
