+++ src/mpi/request/waitany.c
@@ -27,87 +27,6 @@ int MPI_Waitany(int count, MPI_Request array_of_requests[], int *indx, MPI_Statu
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Waitany
 #define MPI_Waitany PMPI_Waitany
-
-/* MPID_Waitany call MPIR_Waitany_state with initialized progress state */
-int MPIR_Waitany_state(int count, MPIR_Request * request_ptrs[], int *indx, MPI_Status * status,
-                       MPID_Progress_state * state)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    for (;;) {
-        int n_inactive = 0;
-        int found_nonnull_req = FALSE;
-
-        for (int i = 0; i < count; i++) {
-            if ((i + 1) % MPIR_CVAR_REQUEST_POLL_FREQ == 0) {
-                mpi_errno = MPID_Progress_test(state);
-                MPIR_ERR_CHECK(mpi_errno);
-            }
-
-            if (request_ptrs[i] == NULL) {
-                ++n_inactive;
-                continue;
-            }
-            /* we found at least one non-null request */
-            found_nonnull_req = TRUE;
-
-            if (MPIR_Request_has_poll_fn(request_ptrs[i])) {
-                mpi_errno = MPIR_Grequest_poll(request_ptrs[i], status);
-                MPIR_ERR_CHECK(mpi_errno);
-            }
-            if (MPIR_Request_is_complete(request_ptrs[i])) {
-                if (MPIR_Request_is_active(request_ptrs[i])) {
-                    *indx = i;
-                    goto fn_exit;
-                } else {
-                    ++n_inactive;
-                    request_ptrs[i] = NULL;
-
-                    if (n_inactive == count) {
-                        *indx = MPI_UNDEFINED;
-                        /* status is set to empty by MPIR_Request_completion_processing */
-                        goto fn_exit;
-                    }
-                }
-            }
-        }
-
-        if (!found_nonnull_req) {
-            /* all requests were NULL */
-            *indx = MPI_UNDEFINED;
-            if (status != NULL) /* could be null if count=0 */
-                MPIR_Status_set_empty(status);
-            goto fn_exit;
-        }
-
-        mpi_errno = MPID_Progress_test(state);
-        MPIR_ERR_CHECK(mpi_errno);
-        /* Avoid blocking other threads since I am inside an infinite loop */
-        MPID_THREAD_CS_YIELD(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    }
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-/* legacy interface (for ch3) */
-int MPIR_Waitany_impl(int count, MPIR_Request * request_ptrs[], int *indx, MPI_Status * status)
-{
-    int mpi_errno = MPI_SUCCESS;
-    MPID_Progress_state progress_state;
-
-    MPID_Progress_start(&progress_state);
-    mpi_errno = MPIR_Waitany_state(count, request_ptrs, indx, status, &progress_state);
-
-    MPID_Progress_end(&progress_state);
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
 #endif
 
 /*@
