+++ src/mpi/request/waitsome.c
@@ -28,84 +28,6 @@ int MPI_Waitsome(int incount, MPI_Request array_of_requests[], int *outcount,
 #ifndef MPICH_MPI_FROM_PMPI
 #undef MPI_Waitsome
 #define MPI_Waitsome PMPI_Waitsome
-
-/* MPID_Waitsome call MPIR_Waitsome_state with initialized progress state */
-int MPIR_Waitsome_state(int incount, MPIR_Request * request_ptrs[],
-                        int *outcount, int array_of_indices[], MPI_Status array_of_statuses[],
-                        MPID_Progress_state * state)
-{
-    int mpi_errno = MPI_SUCCESS;
-
-    /* Bill Gropp says MPI_Waitsome() is expected to try to make
-     * progress even if some requests have already completed;
-     * therefore, we kick the pipes once and then fall into a loop
-     * checking for completion and waiting for progress. */
-    mpi_errno = MPID_Progress_test(state);
-    MPIR_ERR_CHECK(mpi_errno);
-
-    int n_active = 0;
-    for (;;) {
-        int n_inactive = 0;
-
-        for (int i = 0; i < incount; i++) {
-            if ((i + 1) % MPIR_CVAR_REQUEST_POLL_FREQ == 0) {
-                mpi_errno = MPID_Progress_test(state);
-                MPIR_ERR_CHECK(mpi_errno);
-            }
-
-            if (request_ptrs[i] != NULL) {
-                if (MPIR_Request_has_poll_fn(request_ptrs[i])) {
-                    mpi_errno = MPIR_Grequest_poll(request_ptrs[i], &array_of_statuses[i]);
-                    MPIR_ERR_CHECK(mpi_errno);
-                }
-                if (MPIR_Request_is_complete(request_ptrs[i])) {
-                    if (MPIR_Request_is_active(request_ptrs[i])) {
-                        array_of_indices[n_active] = i;
-                        n_active += 1;
-                    } else {
-                        request_ptrs[i] = NULL;
-                        n_inactive += 1;
-                    }
-                }
-            } else {
-                n_inactive += 1;
-            }
-        }
-
-        if (n_active > 0) {
-            *outcount = n_active;
-            break;
-        } else if (n_inactive == incount) {
-            *outcount = MPI_UNDEFINED;
-            break;
-        }
-
-        mpi_errno = MPID_Progress_test(state);
-        MPIR_ERR_CHECK(mpi_errno);
-        /* Avoid blocking other threads since I am inside an infinite loop */
-        MPID_THREAD_CS_YIELD(GLOBAL, MPIR_THREAD_GLOBAL_ALLFUNC_MUTEX);
-    }
-
-  fn_exit:
-    return mpi_errno;
-  fn_fail:
-    goto fn_exit;
-}
-
-/* legacy interface (for ch3) */
-int MPIR_Waitsome_impl(int incount, MPIR_Request * request_ptrs[],
-                       int *outcount, int array_of_indices[], MPI_Status array_of_statuses[])
-{
-    MPID_Progress_state progress_state;
-
-    MPID_Progress_start(&progress_state);
-    int mpi_errno = MPIR_Waitsome_state(incount, request_ptrs, outcount,
-                                        array_of_indices, array_of_statuses, &progress_state);
-    MPID_Progress_end(&progress_state);
-
-    return mpi_errno;
-}
-
 #endif
 
 /*@
